+++ apps/thundermoon/lib/thundermoon/game_of_life/grid.ex
@@ -2,6 +2,7 @@ defmodule Thundermoon.GameOfLife.Grid do
   use GenServer, restart: :temporary
 
   alias Sim.SeamlessGrid, as: Grid
+
   alias ThundermoonWeb.Endpoint
 
   def start_link(size) do